+++ ace-internal.d.ts
@@ -29,6 +29,9 @@ export namespace Ace {
     type Config = typeof import("./src/config");
     type GutterTooltip = import( "./src/mouse/default_gutter_handler").GutterTooltip;
     type GutterKeyboardEvent = import( "./src/keyboard/gutter_handler").GutterKeyboardEvent;
+    type HoverTooltip = import("ace-code/src/tooltip").HoverTooltip;
+    type Tooltip = import("ace-code/src/tooltip").Tooltip;
+    type PopupManager = import("ace-code/src/tooltip").PopupManager;
 
     type AfterLoadCallback = (err: Error | null, module: unknown) => void;
     type LoaderFunction = (moduleName: string, afterLoad: AfterLoadCallback) => void;
@@ -317,7 +320,7 @@ export namespace Ace {
         fullWidth?: boolean,
         screenWidth?: number,
         rowsAbove?: number,
-        lenses?: any[],
+        lenses?: CodeLenseCommand[],
     }
 
     type NewLineMode = 'auto' | 'unix' | 'windows';
@@ -1084,13 +1087,48 @@ export namespace Ace {
         $blockSelectEnabled?: boolean,
     }
 
+    /**
+     * Provider interface for code lens functionality
+     */
     interface CodeLenseProvider {
+        /**
+         * Compute code lenses for the given edit session
+         * @param session The edit session to provide code lenses for
+         * @param callback Callback function that receives errors and code lenses
+         */
         provideCodeLenses: (session: EditSession, callback: (err: any, payload: CodeLense[]) => void) => void;
     }
 
+    /**
+     * Represents a command associated with a code lens
+     */
+    interface CodeLenseCommand {
+        /**
+         * Command identifier that will be executed
+         */
+        id?: string,
+        /**
+         * Display title for the code lens
+         */
+        title: string,
+        /**
+         * Argument(s) to pass to the command when executed
+         */
+        arguments?: any,
+    }
+
+    /**
+     * Represents a code lens - an actionable UI element displayed above a code line
+     */
     interface CodeLense {
+        /**
+         * Starting position where the code lens should be displayed
+         */
         start: Point,
-        command: any
+        /**
+         * Command to execute when the code lens is activated
+         */
+        command?: CodeLenseCommand
     }
 
     interface CodeLenseEditorExtension {