+++ ace.d.ts
@@ -38,6 +38,9 @@ declare module "ace-code" {
         type Config = typeof import("ace-code/src/config");
         type GutterTooltip = import("ace-code/src/mouse/default_gutter_handler").GutterTooltip;
         type GutterKeyboardEvent = import("ace-code/src/keyboard/gutter_handler").GutterKeyboardEvent;
+        type HoverTooltip = import("ace-code/src/tooltip").HoverTooltip;
+        type Tooltip = import("ace-code/src/tooltip").Tooltip;
+        type PopupManager = import("ace-code/src/tooltip").PopupManager;
         type AfterLoadCallback = (err: Error | null, module: unknown) => void;
         type LoaderFunction = (moduleName: string, afterLoad: AfterLoadCallback) => void;
         export interface ConfigOptions {
@@ -222,7 +225,7 @@ declare module "ace-code" {
             fullWidth?: boolean;
             screenWidth?: number;
             rowsAbove?: number;
-            lenses?: any[];
+            lenses?: CodeLenseCommand[];
         }
         type NewLineMode = "auto" | "unix" | "windows";
         interface EditSessionOptions {
@@ -875,12 +878,46 @@ declare module "ace-code" {
             alignCursors: () => void;
             multiSelect?: any;
         }
+        /**
+         * Provider interface for code lens functionality
+         */
         interface CodeLenseProvider {
+            /**
+             * Compute code lenses for the given edit session
+             * @param session The edit session to provide code lenses for
+             * @param callback Callback function that receives errors and code lenses
+             */
             provideCodeLenses: (session: EditSession, callback: (err: any, payload: CodeLense[]) => void) => void;
         }
+        /**
+         * Represents a command associated with a code lens
+         */
+        interface CodeLenseCommand {
+            /**
+             * Command identifier that will be executed
+             */
+            id?: string;
+            /**
+             * Display title for the code lens
+             */
+            title: string;
+            /**
+             * Argument(s) to pass to the command when executed
+             */
+            arguments?: any;
+        }
+        /**
+         * Represents a code lens - an actionable UI element displayed above a code line
+         */
         interface CodeLense {
+            /**
+             * Starting position where the code lens should be displayed
+             */
             start: Point;
-            command: any;
+            /**
+             * Command to execute when the code lens is activated
+             */
+            command?: CodeLenseCommand;
         }
         interface CodeLenseEditorExtension {
             codeLensProviders?: CodeLenseProvider[];
@@ -1007,6 +1044,6 @@ declare module "ace-code" {
     import { Range } from "ace-code/src/range";
     import { UndoManager } from "ace-code/src/undomanager";
     import { VirtualRenderer as Renderer } from "ace-code/src/virtual_renderer";
-    export var version: "1.39.1";
+    export var version: "1.41.0";
     export { Range, Editor, EditSession, UndoManager, Renderer as VirtualRenderer };
 }